Support for Raspberry Pi Pico
Light code retouches, Pi Pico has a perfect microsecond timing so there is no hard coded delays, the Barcode sends the Churro bits by 680 microseconds / bit speed, the Atmega328 has a 3mS +- margin, this is 0 in the Pico. The code is working in Arduino IDE with Pico board extension, tested on Arduino IDE 2.0.2.